requirements.txt
# Database drivers
neo4j>=5.28.0  # Updated to latest version
psycopg2-binary>=2.9.9

# Environment and configuration
python-dotenv>=1.0.0

# File extractors
PyMuPDF>=1.25.5  # Updated to latest version with Mac ARM support
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.0.3
pillow>=10.0.1
pytesseract>=0.3.10
unstructured>=0.11.0

# LLM interaction
ollama>=0.4.7  # Updated to latest version from January 2025
requests>=2.31.0

# Knowledge graph utilities
rdflib>=7.0.0
SPARQLWrapper>=2.0.0

# Analysis tools
networkx>=3.1
scipy>=1.12.0

# Added required packages for wheel building
wheel>=0.42.0
setuptools>=69.0.0